Move accuracy/polars to polars/accuracy + stub
