+++ rust/src/polars/mod.rs
@@ -33,6 +33,9 @@ use polars_plan::{
 #[cfg(feature = "ffi")]
 use serde::{Deserialize, Serialize};
 
+#[cfg(feature = "contrib")]
+pub mod accuracy;
+
 #[cfg(test)]
 mod test;
